code reorganization and optimization
